apps/cli/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concentray-cli"
version = "0.1.0"
description = "CLI for Concentray"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12,<1.0",
  "pydantic>=2.9,<3.0",
  "python-dotenv>=1.0,<2.0",
  "PyYAML>=6.0,<7.0",
  "jsonschema>=4.23,<5.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.7,<1.0"
]

[project.scripts]
concentray = "concentray_cli.main:main"
ctray = "concentray_cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
